refactor: move code into cli project
