--- old/demo/.gitignore
@@ -1,3 +1,0 @@
-Users/
-bundle.js
-bundle.js.map
