--- old/demo/index.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>ng2-nouislider</title>
-    <link
-      rel="stylesheet"
-      href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css"
-    />
-  </head>
-
-  <body>
-    <div class="jumbotron">
-      <div class="container">
-        <h2>ng2-nouislider demos</h2>
-        <a
-          href="https://github.com/tb/ng2-nouislider"
-          class="btn btn-default btn-xs"
-          >Back to GitHub</a
-        >
-        <a href="https://travis-ci.org/tb/ng2-nouislider"
-          ><img src="https://travis-ci.org/tb/ng2-nouislider.svg?branch=master"
-        /></a>
-        <a href="https://badge.fury.io/js/ng2-nouislider"
-          ><img
-            src="https://badge.fury.io/js/ng2-nouislider.svg"
-            alt="npm version"
-            height="18"
-        /></a>
-        <a href="https://npmjs.org/package/ng2-nouislider"
-          ><img
-            src="http://img.shields.io/npm/dm/ng2-nouislider.svg"
-            alt="npm version"
-            height="18"
-        /></a>
-      </div>
-    </div>
-    <div class="container">
-      <app>
-        <br />
-        <br />
-        <p>Loading...</p>
-      </app>
-    </div>
-    <script src="bundle.js"></script>
-  </body>
-</html>
