--- old/demo/polyfills.ts
@@ -1,3 +1,0 @@
-import 'es6-shim';
-import 'reflect-metadata';
-require('zone.js/dist/zone');
