--- old/src/ng2-nouislider.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core';
-import { NouisliderComponent } from './ng2-nouislider.component';
-
-@NgModule({
-  exports: [NouisliderComponent],
-  declarations: [NouisliderComponent],
-})
-export class NouisliderModule {}
