--- old/src/public_api.ts
@@ -1,2 +1,0 @@
-export * from './ng2-nouislider.component';
-export * from './ng2-nouislider.module';
