+++ projects/demo/src/app/app.component.scss
@@ -1,0 +1,32 @@
+@import '~nouislider/distribute/nouislider.min.css';
+
+body {
+  padding-bottom: 100px;
+}
+
+code {
+  font-size: 1.5rem;
+}
+
+pre,
+[nouislider] {
+  margin: 15px 0;
+}
+
+.jumbotron {
+  background: #f5f5f5;
+  border-bottom: 1px solid #eee;
+  color: #333;
+  margin-bottom: 0px;
+  padding-top: 0px;
+  padding-bottom: 15px;
+}
+
+.noUi-connect {
+  background: #337ab7;
+}
+
+.event-label {
+  transition: all 0.25s;
+  color: white !important;
+}
