+++ projects/demo/src/app/app.module.ts
@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { NouisliderModule } from 'ng2-nouislider';
 
 import { AppComponent } from './app.component';
 
 @NgModule({
+  imports: [BrowserModule, FormsModule, NouisliderModule, ReactiveFormsModule],
   declarations: [AppComponent],
-  imports: [BrowserModule],
   providers: [],
   bootstrap: [AppComponent],
 })
