+++ projects/demo/src/index.html
@@ -2,12 +2,47 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <title>Demo</title>
+    <title>ng2-nouislider</title>
     <base href="/" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="icon" type="image/x-icon" href="favicon.ico" />
+    <link
+      rel="stylesheet"
+      href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css"
+    />
   </head>
   <body>
-    <app-root></app-root>
+    <div class="jumbotron">
+      <div class="container">
+        <h2>ng2-nouislider demos</h2>
+        <a
+          href="https://github.com/tb/ng2-nouislider"
+          class="btn btn-default btn-xs"
+          >Back to GitHub</a
+        >
+        <a href="https://travis-ci.org/tb/ng2-nouislider"
+          ><img src="https://travis-ci.org/tb/ng2-nouislider.svg?branch=master"
+        /></a>
+        <a href="https://badge.fury.io/js/ng2-nouislider"
+          ><img
+            src="https://badge.fury.io/js/ng2-nouislider.svg"
+            alt="npm version"
+            height="18"
+        /></a>
+        <a href="https://npmjs.org/package/ng2-nouislider"
+          ><img
+            src="http://img.shields.io/npm/dm/ng2-nouislider.svg"
+            alt="npm version"
+            height="18"
+        /></a>
+      </div>
+    </div>
+    <div class="container">
+      <app-root>
+        <br />
+        <br />
+        <p>Loading...</p>
+      </app-root>
+    </div>
   </body>
 </html>
