+++ projects/ng2-nouislider/src/lib/ng2-nouislider.module.ts
@@ -1,9 +1,8 @@
 import { NgModule } from '@angular/core';
-import { Ng2NouisliderComponent } from './ng2-nouislider.component';
+import { NouisliderComponent } from './ng2-nouislider.component';
 
 @NgModule({
-  declarations: [Ng2NouisliderComponent],
-  imports: [],
-  exports: [Ng2NouisliderComponent],
+  exports: [NouisliderComponent],
+  declarations: [NouisliderComponent],
 })
-export class Ng2NouisliderModule {}
+export class NouisliderModule {}
