--- projects/ng2-nouislider/src/lib/ng2-nouislider.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { Ng2NouisliderService } from './ng2-nouislider.service';
-
-describe('Ng2NouisliderService', () => {
-  let service: Ng2NouisliderService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(Ng2NouisliderService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
