--- projects/ng2-nouislider/src/lib/ng2-nouislider.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class Ng2NouisliderService {
-  constructor() {}
-}
