+++ projects/ng2-nouislider/src/public-api.ts
@@ -2,6 +2,5 @@
  * Public API Surface of ng2-nouislider
  */
 
-export * from './lib/ng2-nouislider.service';
 export * from './lib/ng2-nouislider.component';
 export * from './lib/ng2-nouislider.module';
